% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{add_timer_record}
\alias{add_timer_record}
\title{Add a timer record}
\usage{
add_timer_record(project_id, start, end, body = "")
}
\arguments{
\item{project_id}{Integer - the ID for the project to assign this `timer` record to}

\item{start}{POSIXct datetime - the start date and time for this period of activity}

\item{end}{POSIXct datetime - the end date and time for this period of activity}

\item{body}{String (optional) - a description of the work carried out}
}
\value{
NA
}
\description{
Adds a record to the `timer` database table to record a period of time
working on a project.
}
\details{
This function handles the insertion of a new record to the `timer` database.
Supply the details for which project this activity was conducted for, the 
start and end datetimes and, optionally, a description of what work was
done, and the function will insert it to the `timers` table of the `timer`
database.
}
\examples{
add_timer_record(
 project_id = 1,
 start = as.POSIXct("2025-08-12 12:00:00"),
 end = as.POSIXct("2025-08-12 13:47:13"),
 body = "Working the {timer} database app"
)
}
