% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarm.r
\name{alarm}
\alias{alarm}
\title{alarm}
\usage{
alarm(timestr)
}
\arguments{
\item{timestr}{Time(s) that the alarm will go off.}
}
\description{
A count down until the clock matches the given time(s).
}
\examples{
\dontrun{
alarm("01:32")
alarm("14:59:15")
alarm(c("01:32","14:59:15"))
}

}
