% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{auth_header}
\alias{auth_header}
\title{Get encoded auth header}
\usage{
auth_header(token)
}
\arguments{
\item{token}{access token for desired user.}
}
\value{
encoded header value suitable for use in httr::add_headers
}
\description{
Convenience wrapper format authorization header value based on token
}
