% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{bench}
\alias{bench}
\title{Computation benchmarking}
\usage{
bench(scheme, ...)
}
\arguments{
\item{scheme}{One of "large", "medium" or "small" describing the scale of
the benchmark. This scales the number of tests as well as the
amount of data used in some of them.}

\item{...}{Provide additional arbitrary functions to test}
}
\description{
Computation benchmarking
}
