% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{calculate_duration}
\alias{calculate_duration}
\title{Calculate the duration between two time periods}
\usage{
calculate_duration(.start_time, .end_time)
}
\arguments{
\item{.start_time}{POSIXct datetime - the start of the time period}

\item{.end_time}{POSIXct datetime - the end of the time period}
}
\value{
Numeric, double - a lubridate `period` object
}
\description{
Calculates the duration between two time periods.
}
\details{
Uses functions from the lubridate package to first calculate the `interval`,
which is defined as 'time spans bound by two real date-times', which are
rounded to the nearest second, before being converted to a `period` to
report the elapsed time using the appropriate units, e.g. 1H 18M 0S.
}
\examples{
calculate_duration(
  .start_time = lubridate::as_datetime("2025-08-13 07:16:00"),
  .end_time = lubridate::as_datetime("2025-08-13 08:34:00")
)
}
