% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{case}
\alias{case}
\title{A \code{case}-like function for R}
\usage{
case(..., otherwise = NULL)
}
\arguments{
\item{...}{arguments to be passed to the function}

\item{otherwise}{the return value in the case that no condition matches}
}
\description{
Given a condition, evaluate cases and return the requested value for the
first condition to be satisfied. If no case matches, return \code{NULL}
}
