% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_datetime}
\alias{check_datetime}
\title{Check whether a supplied datetime is of POSIXct type}
\usage{
check_datetime(datetime_value)
}
\arguments{
\item{datetime_value}{Input expected to be of POSIXct type}
}
\value{
POSIXct datetime_value or an error raised
}
\description{
If the supplied value inherits from the 'POSIXct' data type then it is
returned, otherwise the function call is aborted and console error raised.
}
\examples{
# if the parameter is valid POSIXct then it is returned
check_datetime(as.POSIXct("2025-08-12 08:00:00"))

# if the parameter is not a valid POSIXct value then an error is returned
check_datetime("2025-08-12 08:00:00")
}
