% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_controller.R
\name{check_project_id}
\alias{check_project_id}
\title{Checks for a valid `project_id`}
\usage{
check_project_id(input_project_id)
}
\arguments{
\item{.input_project_id}{Numeric - input value to check if exists}
}
\value{
Numeric - the ID value if it exists as a project_id, otherwise raises an error
}
\description{
Checks whether the supplied `project_id` exists within the `projects` table.
}
