% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockify.R
\name{clockify_get}
\alias{clockify_get}
\title{GET requests to clockify endpoints}
\usage{
clockify_get(endpoint, token)
}
\arguments{
\item{endpoint}{desired endpoint to GET (e.g. "/workspaces")}

\item{token}{access token for desired user. If null, will try to load from
environment variable CLOCKIFY_TOKEN}
}
\value{
response code returned by request
}
\description{
Convenience wrapper to fetch data from clockify API GET endpoints
}
