% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockify.R
\name{clockify_get_running_task}
\alias{clockify_get_running_task}
\title{Get running task}
\usage{
clockify_get_running_task(workspace_id, token)
}
\arguments{
\item{workspace_id}{id for the desired workspace}

\item{token}{access token for desired user.}
}
\value{
Named list with
}
\description{
Get running task
}
