% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockify.R
\name{clockify_get_workspaces}
\alias{clockify_get_workspaces}
\title{Get workspaces}
\usage{
clockify_get_workspaces(token = NULL)
}
\arguments{
\item{token}{access token for desired user.}
}
\value{
List of workspaces, with name and id for each workspace
}
\description{
Get workspaces
}
