% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockify.R
\name{clockify_start_timer}
\alias{clockify_start_timer}
\title{Start the timer}
\usage{
clockify_start_timer(workspace_id, project_id, description, token = NULL)
}
\arguments{
\item{workspace_id}{id for the desired workspace}

\item{project_id}{id for the desired project}

\item{description}{description for work being done}

\item{token}{access token for desired user.}
}
\value{
ID for started time entry
}
\description{
Start the timer
}
