% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockify.R
\name{clockify_stop_timer}
\alias{clockify_stop_timer}
\title{Stop the timer}
\usage{
clockify_stop_timer(workspace_id, token = NULL)
}
\arguments{
\item{workspace_id}{id for the desired workspace}

\item{token}{access token for desired user.}
}
\value{
result of API call
}
\description{
Stop the timer
}
