% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockify.R
\name{clockify_update_buttons}
\alias{clockify_update_buttons}
\title{Toggle button enabled status}
\usage{
clockify_update_buttons(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
None. Called for side effect of updating UI elements
}
\description{
Toggle button enabled status
}
