% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_bar}
\alias{create_bar}
\title{Writing a fixed width bar}
\usage{
create_bar(content = NULL, width = 80)
}
\arguments{
\item{content}{extra text to include in the bar}

\item{width}{how far to draw the line}
}
\description{
Return either a solid horizontal box character or specified text surrounded
by horizontal box characters for a specified width.
}
\examples{
create_bar()  # write just a bar
create_bar("Content in here")
}
