% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_controller.R
\name{create_db}
\alias{create_db}
\title{Create and initialise the timer database}
\usage{
create_db()
}
\value{
NA
}
\description{
Creates a {SQLite} database to hold data for the {timeR} application 
and initialised for use.
}
\details{
If the database does not exist, then it will be created. 

If the `projects` table does not exist, then it will be created and
initialised with common projects, such as 'Non-chargeable time', 'Annual
leave' and so on.

If the `timers` table does not exist, then it will be created and indexes
created on both the `id_date` and `id_week` columns. As their name indicates
these columns are identifier columns used to uniquely identify a day (in 
YYYYMMDD notation) and any given week (using YYYYWW notation). These columns
are calculated fields and are designed to improve query responsiveness.
}
\examples{
# This function is designed to be called at the beginning of the app, doing
# nothing more than checking the necessary database and tables are set up.
create_db()
}
