% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer_function.R
\name{createTimer}
\alias{createTimer}
\title{Create a timer object}
\usage{
createTimer(verbose = T)
}
\arguments{
\item{verbose}{A parameter to control whether to print messages while using
methods. Default to \code{TRUE}.}
}
\value{
a timer object.
}
\description{
Create a timer object
}
\examples{
timer1 <- createtimer() #print is enabled
timer <- createtimer(F) #print is disabled
}
