% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{get_db_path}
\alias{get_db_path}
\title{Get the database file path}
\usage{
get_db_path()
}
\value{
String - file path to the database file
}
\description{
Returns the file path to the database
}
\details{
An internal function to return the file path to the database, called
`timer_db.sqlite` which is stored within the `inst` folder.
}
