% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_controller.R
\name{get_id_date}
\alias{get_id_date}
\title{Get a unique ID for the date}
\usage{
get_id_date(datetime)
}
\arguments{
\item{datetime}{POSIXct datetime}
}
\value{
Numeric - an eight digit number used as an id number for the date
}
\description{
Generates a numeric ID from a POSIXct datetime that uniquely identifies 
the date
}
\details{
Returns an integer which can be used as a unique identifier for the date
encoded in the 'datetime' object. 

It works by first formatting the datetime to a string with the format 
YYYYMMDD then casting to numeric type.
}
\examples{
get_id_date(as.POSIXct("2025-08-12 12:00:00"))
get_id_date(lubridate::now())
}
