% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{get_id_week}
\alias{get_id_week}
\title{Get a unique ID for the week}
\usage{
get_id_week(datetime)
}
\arguments{
\item{datetime}{POSIXct datetime}
}
\value{
Numeric - a six digit number used as an id for the week
}
\description{
Generates a numeric ID from a POSIXct datetime that uniquely identifies 
the week
}
\details{
Returns an integer which can be used as a unique identifier for the week
encoded in the 'datetime' object.

It uses a combination of the ISO-year and ISO-week from the `lubridate`
package to generate an integer that uniquely identifies the week 
across years.

It does this by generating the ISO-year, e.g 2025, then multiplying by
100 to move the digits two decimal places, e.g. 202500. Then it adds the
ISO week number, e.g. 33 to this to generate an integer, e.g. 202533 that
is unique across years.
}
\examples{
get_id_week(as.POSIXct("2025-08-12 12:00:00"))
get_id_week(lubridate::now())
}
