% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_project_choices}
\alias{get_project_choices}
\title{Get details for projects ready for use in a select input}
\usage{
get_project_choices()
}
\value{
List of named projects with the corresponding database id numbers
}
\description{
Returns a named list of projects ready for use in a select input
}
\details{
Gets a tibble of project details from the `get_project_records()` function,
limits the details to project name and id values then creates a named list
suitable for use with a select input control.
}
\examples{
get_project_choices()
}
\seealso{
[get_project_records()] To return a tibble of project details
}
