% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{get_project_records}
\alias{get_project_records}
\title{Get `projects` records}
\usage{
get_project_records(.id_project = NA)
}
\arguments{
\item{.id_project}{Numeric - an ID value that uniquely identifies a project}
}
\value{
Tibble of project records
}
\description{
Returns records from the `projects` table that match the supplied id for the
project, or all records if the parameter is empty.
}
\details{
Conducts a SELECT query from the `projects` table to return records. If a 
value is supplied for `.id_project` then the query returns the record for
that project.
}
\examples{
# calling the function with no arguments will return all records
get_project_records()

# calling the function with an .id_project will return the record for that
# project
get_project_records(.id_project = 1)
---
}
