% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{get_project_records_for_calendar}
\alias{get_project_records_for_calendar}
\title{Get `projects` records for a calendar}
\usage{
get_project_records_for_calendar(.id_project = NA)
}
\arguments{
\item{.id_project}{Numeric - an ID value that uniquely identifies a project}
}
\value{
Tibble of `projects` records
}
\description{
Returns a tibble of formatted data ready for use with the calendar display
}
\details{
The `toastui::calendar()` works best with details in string format, so one 
of the tasks of this function is to loads project details from the database
and convert each variable to character(). Another task is to prepare the
colour scheme for display in `toastui::calendar()` by taking the hex colour
`colour` and using it to produce a colour scheme for `backgroundColor`,
a semi-transparent colour, `borderColor` a solid colour used  and `color`
used for the font.
}
\examples{
# get all project records
get_project_records_for_calendar()

# get the record for a specific project id
get_project_records_for_calendar(.id_project = 1)
}
\seealso{
[get_project_records()] for the underlying function to query projects from
the database.
}
