% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_controller.R
\name{get_timer_records}
\alias{get_timer_records}
\title{Get `timers` records}
\usage{
get_timer_records(.id_date = NA, .id_week = NA)
}
\arguments{
\item{.id_date}{Numeric - an ID value that uniquely identifies a date, in the format 'YYYYMMDD',}

\item{.id_week}{Numeric - an ID value that uniquely identifies a week, in the format 'YYYYWW'}
}
\value{
Tibble
}
\description{
Returns records from the `timers` table that match the supplied id value for 
the date or week, or all records if both parameters are empty.
}
\details{
Conducts a SELECT query from the `timers` table to return records. If values
are supplied for either or both `.id_date` or `.id_week` then the query 
returns records for that day or week respectively.

Please note, if both parameters are supplied the query will return the 
intersection of the two, so if the date for `.id_date` did not fall within
the range of dates for `.id_week` then no results will be returned.
}
\examples{
# calling the function with no arguments will return all records
get_timer_records()

# calling the function with either .id_date or .id_week will return records for
# that date or week, respectively
get_timer_records(.id_date = 20250812)
get_timer_records(.id_week = 202533)

# if you don't know the id number but have a POSIXct value you can combine it
# with the `get_id_date()` or `get_id_week()` helper functions
get_timer_records(
  .id_week = get_id_week(lubridate::as.datetime("2025-08-13 08:50:00"))
)
}
\seealso{
[get_id_date()] for a function that returns an .id_date for a given POSIXct datetime, and
[get_id_week()] for a function that reuturns an .id_week for a given POSIXct datetime.
[get_project_records()] does a similar function for project records.
}
