% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{get_timer_records_for_calendar}
\alias{get_timer_records_for_calendar}
\title{Get `timers` records for a calendar}
\usage{
get_timer_records_for_calendar(.id_date = NA, .id_week = NA)
}
\arguments{
\item{.id_date}{Numeric - an ID value that uniquely identifies a date, in the format 'YYYYMMDD',}

\item{.id_week}{Numeric - an ID value that uniquely identifies a week, in the format 'YYYYWW'}
}
\value{
Tibble of `timers` records
}
\description{
Returns a tibble of formatted data ready for use with the calendar display
}
\details{
The `toastui::calendar()` works best with details in string format, so one 
of the tasks of this function is to loads timer and project details from 
the database and convert each variable to character(). Another task is to 
add the name of the project as the 'title' for the calendar entry, as this
is a required field by `toastui::calendar()`.
}
\examples{
# return all timer records
get_timer_records_for_calendar()

# return the timer records for this week
get_timer_records_for_calendar(.id_week = get_id_week(lubridate::now()))

# return the timer records for today
get_timer_records_for_calendar(.id_day = get_id_day(lubridate::now()))
}
\seealso{
[get_timer_records()] for the underlying function to query timers from the
the database.

[get_project_records()] for the underlying function to query projects from
the database.
}
