% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database.R
\name{get_timer_records_for_table}
\alias{get_timer_records_for_table}
\title{Get `timers` records for a table}
\usage{
get_timer_records_for_table(.id_date = NA, .id_week = NA)
}
\arguments{
\item{.id_date}{Numeric - an ID value that uniquely identifies a date, in the format 'YYYYMMDD',}

\item{.id_week}{Numeric - an ID value that uniquely identifies a week, in the format 'YYYYWW'}
}
\value{
Tibble of `timers` records
}
\description{
Returns a tibble of formatted data ready for use with the datagrid display
}
\details{
The `toastui::datagrid()` is used to display a list of timer records. This
function loads the data and prepares it for display by converting `start`
and `end` variables to timestamps.
}
\examples{
# return all timer records
get_timer_records_for_table()

# return the timer records for this week
get_timer_records_for_table(.id_week = get_id_week(lubridate::now()))

# return the timer records for today
get_timer_records_for_table(.id_day = get_id_day(lubridate::now()))
}
\seealso{
[get_timer_records()] for the underlying function to query timers from the
the database.

[get_project_records()] for the underlying function to query projects from
the database.
}
