% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_timescales_list}
\alias{get_timescales_list}
\title{Get a list of timescales}
\usage{
get_timescales_list()
}
\value{
Named list of time periods and corresponding 'id' values
}
\description{
Returns a named list of timescales along with the corresponding id value
for the day or week for use with the 'data view' module
}
\details{
Returns a named list of time periods:
- 'Today' and 'Yesterday' contain the id values for the corresponding days
- 'This week' and 'Last week' contain the id values for the corresponding weeks
- 'All' is `NA`.
}
\examples{
get_timescales_list()
}
