% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer_function.R
\name{getTimer}
\alias{getTimer}
\title{Get the data frame in timer object}
\usage{
getTimer(object)
}
\arguments{
\item{object}{The name for timer object.}
}
\value{
A data frame containing all records of a timer object.
}
\description{
timer object has a built-in data frame that contains all timings. run this
function to extract the data frame.
}
\examples{
timer1 <- createTimer()
timer1$start("event1")
Sys.sleep(1)
timer1$stop("event1")
getTimer(timer1)
}
