% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{is.timer}
\alias{is.timer}
\title{Is Timer}
\usage{
is.timer(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class \code{\link{Timer}}.
}

