% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nprof-package.R
\docType{package}
\name{timeR-package}
\alias{timeR}
\alias{timeR-package}
\title{timeR: R functions to test system performance}
\description{
A suite of functions for benchmarking hardware performance.
}
\author{
\strong{Maintainer}: Gareth Clews \email{20670199+karetsu@users.noreply.github.com}

}
\keyword{internal}
