% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_get}
\alias{toggl_get}
\title{GET requests to toggl endpoints}
\usage{
toggl_get(endpoint, token)
}
\arguments{
\item{endpoint}{desired endpoint to GET (e.g. "/me")}

\item{token}{access token for desired user. If null, will try to load from
environment variable TOGGL_TOKEN}
}
\value{
response code returned by request
}
\description{
Convenience wrapper to fetch data from toggle API GET endpoints
}
