% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_get_projects}
\alias{toggl_get_projects}
\title{Get projects}
\usage{
toggl_get_projects(workspace_id, token = NULL)
}
\arguments{
\item{workspace_id}{id for the workspace from which to load projects}

\item{token}{access token for desired user.}
}
\value{
List of projects, with name and id for each workspace
}
\description{
Get projects
}
