% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_get_running_task}
\alias{toggl_get_running_task}
\title{Get running task}
\usage{
toggl_get_running_task(token)
}
\arguments{
\item{token}{access token for desired user.}
}
\value{
Named list with
}
\description{
Get running task
}
