% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_is_running}
\alias{toggl_is_running}
\title{Is a time entry currently running}
\usage{
toggl_is_running(token)
}
\arguments{
\item{token}{access token for desired user.}
}
\value{
Boolean indicating if a time entry is currently running
}
\description{
Is a time entry currently running
}
