% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_post}
\alias{toggl_post}
\title{POST/PUT requests to toggl endpoints}
\usage{
toggl_post(endpoint, data, token, method = "POST")
}
\arguments{
\item{endpoint}{desired endpoint to GET (e.g. "/workspaces")}

\item{data}{data to submit with post}

\item{token}{access token for desired user.}

\item{method}{POST by default, but call be used for "PUT".}
}
\value{
response code returned by request
}
\description{
Convenience wrapper to fetch data from toggle API POST/PUT endpoints
}
