% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_start_timer}
\alias{toggl_start_timer}
\title{Start the timer}
\usage{
toggl_start_timer(project_id, description, token = NULL)
}
\arguments{
\item{project_id}{id for the desired project}

\item{description}{description for work being done}

\item{token}{access token for desired user.}
}
\value{
ID for started time entry
}
\description{
Start the timer
}
