% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_stop_timer}
\alias{toggl_stop_timer}
\title{Stop the timer}
\usage{
toggl_stop_timer(token = NULL)
}
\arguments{
\item{token}{access token for desired user.}

\item{time_entry_id}{id for the desired workspace}
}
\value{
result of API call
}
\description{
Stop the timer
}
