% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webtimer.r
\name{webtimer}
\alias{webtimer}
\title{webtimer}
\usage{
webtimer(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Should the app automatically launch in browser?}
}
\description{
Runs a shiny-based webapp timer.
}
\details{
If shiny is not available, then this does nothing.
}
