% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer_function.R
\name{createTimer}
\alias{createTimer}
\title{Create a timer object}
\usage{
createTimer(verbose = T)
}
\arguments{
\item{verbose}{A parameter to control whether to print messages while using
methods. Default to \code{TRUE}.}
}
\value{
a timer object.
}
\description{
Create a timer object
}
\examples{
timer1 <- createTimer() # print is enabled
timer1 <- createTimer(F) # print is disabled
timer1$start("event1") # start timing for event 1
timer1$stop("event1", comment = "event 1 stopped") # stop timing for event 1(comment is optional)
getTimer(timer1) # get all records in a data frame
}
