% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer-package.R
\docType{package}
\name{timer-package}
\alias{timer}
\alias{timer-package}
\title{timer: R6 Timer}
\description{
A simple timer as an R6 class. The timer has four functions: start(), stop(), reset() and elapsed(). The elapsed() function returns the elapsed wall clock time (as opposed to CPU time) as an object of class lubridate::Duration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/timer}
  \item \url{https://poissonconsulting.github.io/timer/}
  \item Report bugs at \url{https://github.com/poissonconsulting/timer/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [funder, copyright holder]
}

}
\keyword{internal}
