% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.r
\name{timer}
\alias{timer}
\alias{hours}
\alias{minutes}
\alias{seconds}
\title{timer}
\usage{
timer(hours = 0, minutes = 0, seconds = 0)

hours(hours)

minutes(minutes)

seconds(seconds)
}
\arguments{
\item{hours, minutes, seconds}{Time to count down from.}
}
\description{
A count down timer.
}
