% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.r
\name{timer}
\alias{timer}
\alias{hours}
\alias{minutes}
\alias{seconds}
\title{timer}
\usage{
timer(hours = 0, minutes = 0, seconds = 0, recur = FALSE)

hours(hours, recur = FALSE)

minutes(minutes, recur = FALSE)

seconds(seconds, recur = FALSE)
}
\arguments{
\item{hours, minutes, seconds}{Time to count down from.}

\item{recur}{Should the timer repeat?}
}
\description{
A count down timer.
}
