% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R, R/timer_class.R
\docType{package}
\name{timeR}
\alias{timeR}
\alias{timeR-package}
\title{timeR: A package to make timing codes easier}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
timeR
}
\description{
The timeR package saves your time by timing your code and save
recordings to a data frame automatically.So you don't have to do all these
steps manually by yourself.

timer is a R6 Class that represent a timer.
}
\section{Fields}{

\describe{
\item{\code{time}}{A POSIXct/POSIXlt value of your latest timing.}

\item{\code{event}}{A string of your latest timing.}

\item{\code{eventTable}}{A data frame that stores all timings.}

\item{\code{verbose}}{A printing setting that controls whether to print messages.}
}}

\section{Public Methods}{

\describe{
  \item{\code{initialize(time,event,verbose,eventTable)}
  }{Initialize a timer object. You can also use \code{createTimer()}
  function to initialize a timer object.}
  \item{\code{start(eventName)}
  }{Start timing for a event, \code{eventName} should be a string}
  \item{\code{stop(eventName)}
  }{Stop timing for a event.}
  \item{\code{getTimer()}
  }{Get a data frame that stores all recordings.You can also use
  \code{getTimer()} function to get the data frame.}
  \item{\code{removeEvent(eventName)}
  }{Remove an given row in the eventTable.}
  \item{\code{toggleVerbose()}
  }{Toggle between \code{TRUE} and \code{FALSE} for \code{verbose}}
  \item{\code{getStartTime()}
  }{Get start time for a selected event.}
  \item{\code{getStopTime()}
  }{Get stop time for a selected event.}
  \item{\code{getTimeElapsed()}
  }{Get time elapsed for a selected event.}
  \item{\code{getComment()}
  }{Get comment for a selected event.}
  \item{\code{print()}
  }{Custom print method for timer class. However, you don't need to use this
  function to generate custom printing.
  Custom printing is triggered by default.}
  }
}

\section{Private Methods}{

\describe{
\item{\code{slprint(msg,flag = self$verbose)}
}{A function that controls whether to print extra message.}
}
}

\examples{
timer <- createTimer()
timer$start("event1")
# put some codes in between
timer$stop("event1")

timer$start("event2")
# put some codes in between
timer$stop("event2",comment = "event 2 completed")

table1 <- getTimer(timer)
timer$toggleVerbose() # set verbose to FALSE as default is TRUE

table1 # print all records in a tibble(data frame)

# get attributes for selected events
timer$getStartTime("event 1")
timer$getStopTime("event 1")
timer$getTimeElapsed("event 1")
timer$getComment("event 1")
}
\keyword{datasets}
